+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Refactor: merge templates for sidebars and example blocks.
 - Consolidate template and refine the style of listing and litertal blocks.
 - Adjust presentation of inline code.
+- Custom AsciiDoc templates for processing all in-use AsciiDoc block types.
 
 -----